Cleanup r_anal_function_contains()
